Fix scoped associations with empty loaded records
This impacts associations with a scope like so:
```
class Coupon < ActiveRecord::Base
  has_many :coupon_redemptions, -> { where(expired: false) }
end
```

Using the example above, let's say we hold a loaded `coupon_redemption`
record in memory and it switches from `expired: true` to `expired:
false`. Assuming that change is persisted to the database, we should
respect it when working with that record in memory.

That means `coupon.coupon_redemptions.empty?` should now be true.

Co-authored-by: Daniel Colson <composerinteralia@github.com>