Update primary key for dl_keyed_join
The primary_key for the `dl_keyed_join` relation was incorrect. It
expected to use  a `joins_key` attribute on the parent record in order
to link to the `dl_keyed_join` record. The parent class does not have a
`joins_key` attribute at all, which means that association was never
setup correctly.

Co-authored-by: Daniel Colson <composerinteralia@github.com>