Clear out target when no new records
When counting records for a has_many association, we were previously
getting in an out-of-sync state if there were records in memory that
were persisted to the database, but did not count for the scoped query.

This impacts associations with a scope like so:
```
class Coupon < ActiveRecord::Base
  has_many :coupon_redemptions, -> { where(expired: false) }
end
```

Using the example above, let's say we hold a loaded `coupon_redemption`
record in memory and it switches from `expired: true` to `expired:
false`. Assuming that change is persisted to the database, we should
respect it when working with that record in memory.

That means `coupon.coupon_redemptions.size` should now be 0. That's not
what were seeing earlier.

The fix was to ensure the has_many association target list is empty when
counting records in that scenario.

Co-authored-by: Daniel Colson <composerinteralia@github.com>